Added a passthrough password encoder
This allows the underlying infrastructure to perform the password
encoding and just performs a string comparison for the passwords. This
was necessary in order to allow clients to be created through multiple
mechanisms in VMidentity (e.g. REST IDM or UAA).

One unfortunate side-effect of this change is that the password encoding
is inherently reversible, but this is already the case and regularly
used for Identity Source passwords when stored in VMDir.